requirements.txt
fastapi>=0.111,<1.0
uvicorn[standard]>=0.30,<1.0
pydantic>=2.6,<3.0
jsonschema>=4.21,<5.0
httpx>=0.27,<1.0
orjson>=3.10,<4.0
pyyaml>=6.0,<7.0
cryptography>=42.0,<44.0
redis>=5.0,<6.0
asyncpg>=0.29,<0.30
google-cloud-firestore>=2.15,<3.0
